Move DB services to a sub-directory
